Lots of inside-of-package code reorg, having absolutely no impact on the behavior or scope of any code.
