+++ pipeline/filters.go
@@ -18,49 +18,8 @@ package pipeline
 import (
 	"fmt"
 	"sort"
-	"sync"
 	"time"
 )
-
-// Heka PluginRunner interface for Filter type plugins.
-type FilterRunner interface {
-	PluginRunner
-	// Input channel on which the Filter should listen for incoming messages
-	// to be processed. Closure of the channel signals shutdown to the filter.
-	InChan() chan *PipelinePack
-	// Associated Filter plugin object.
-	Filter() Filter
-	// Starts the Filter (so it's listening on the input channel for messages
-	// to be processed) in a separate goroutine and returns. Should decrement
-	// the wait group when the Filter has stopped and the goroutine has
-	// completed.
-	Start(h PluginHelper, wg *sync.WaitGroup) (err error)
-	// Returns a ticker channel configured to send ticks at an interval
-	// specified by the plugin's ticker_interval config value, if provided.
-	Ticker() (ticker <-chan time.Time)
-	// Hands provided PipelinePack to the Heka Router for delivery to any
-	// Filter or Output plugins with a corresponding message_matcher. Returns
-	// false and doesn't perform message injection if the message would be
-	// caught by the sending Filter's message_matcher.
-	Inject(pack *PipelinePack) bool
-	// Parsing engine for this Filter's message_matcher.
-	MatchRunner() *MatchRunner
-	// Retains a pack for future delivery to the plugin when a plugin needs
-	// to shut down and wants to retain the pack for the next time its
-	// running properly
-	RetainPack(pack *PipelinePack)
-}
-
-// Heka Filter plugin type.
-type Filter interface {
-	// Starts the filter listening on the FilterRunner's provided input
-	// channel. Should not return until shutdown, signaled to the Filter by
-	// the closure of the input channel. Should return a non-nil error value
-	// only if errors happen during start-up or if there is an unclean
-	// shutdown (i.e. not due to an error processing an isolated message, in
-	// that case use FilterRunner.LogError).
-	Run(r FilterRunner, h PluginHelper) (err error)
-}
 
 // Filter that counts the number of messages flowing through and provides
 // primitive aggregation counts.
