+++ pipeline/inputs.go
@@ -35,165 +35,6 @@ import (
 	"time"
 )
 
-// Heka PluginRunner for Input plugins.
-type InputRunner interface {
-	PluginRunner
-	// Input channel from which Inputs can get fresh PipelinePacks, ready to
-	// be populated.
-	InChan() chan *PipelinePack
-	// Associated Input plugin object.
-	Input() Input
-
-	SetTickLength(tickLength time.Duration)
-
-	// Returns a ticker channel configured to send ticks at an interval
-	// specified by the plugin's ticker_interval config value, if provided.
-	Ticker() (ticker <-chan time.Time)
-
-	// Starts Input in a separate goroutine and returns. Should decrement the
-	// plugin when the Input stops and the goroutine has completed.
-	Start(h PluginHelper, wg *sync.WaitGroup) (err error)
-	// Injects PipelinePack into the Heka Router's input channel for delivery
-	// to all Filter and Output plugins with corresponding message_matchers.
-	Inject(pack *PipelinePack)
-}
-
-type iRunner struct {
-	pRunnerBase
-	input      Input
-	inChan     chan *PipelinePack
-	tickLength time.Duration
-	ticker     <-chan time.Time
-}
-
-func (ir *iRunner) SetTickLength(tickLength time.Duration) {
-	ir.tickLength = tickLength
-}
-
-func (ir *iRunner) Ticker() (ticker <-chan time.Time) {
-	return ir.ticker
-}
-
-// Creates and returns a new (not yet started) InputRunner associated w/ the
-// provided Input.
-func NewInputRunner(name string, input Input, pluginGlobals *PluginGlobals) (
-	ir InputRunner) {
-	return &iRunner{
-		pRunnerBase: pRunnerBase{
-			name:          name,
-			plugin:        input.(Plugin),
-			pluginGlobals: pluginGlobals,
-		},
-		input: input,
-	}
-}
-
-func (ir *iRunner) Input() Input {
-	return ir.plugin.(Input)
-}
-
-func (ir *iRunner) InChan() chan *PipelinePack {
-	return ir.inChan
-}
-
-func (ir *iRunner) Start(h PluginHelper, wg *sync.WaitGroup) (err error) {
-	ir.h = h
-	ir.inChan = h.PipelineConfig().inputRecycleChan
-
-	if ir.tickLength != 0 {
-		ir.ticker = time.Tick(ir.tickLength)
-	}
-
-	go ir.Starter(h, wg)
-	return
-}
-
-func (ir *iRunner) Starter(h PluginHelper, wg *sync.WaitGroup) {
-	defer func() {
-		wg.Done()
-	}()
-
-	globals := Globals()
-	rh, err := NewRetryHelper(ir.pluginGlobals.Retries)
-	if err != nil {
-		ir.LogError(err)
-		globals.ShutDown()
-		return
-	}
-
-	for !globals.Stopping {
-		// ir.Input().Run() shouldn't return unless error or shutdown
-		if err := ir.Input().Run(ir, h); err != nil {
-			ir.LogError(err)
-		} else {
-			ir.LogMessage("stopped")
-		}
-
-		// Are we supposed to stop? Save ourselves some time by exiting now
-		if globals.Stopping {
-			return
-		}
-
-		// We stop and let this quit if its not a restarting plugin
-		if recon, ok := ir.plugin.(Restarting); ok {
-			recon.CleanupForRestart()
-		} else {
-			ir.LogMessage("has stopped, shutting down.")
-			globals.ShutDown()
-			return
-		}
-
-		// Re-initialize our plugin using its wrapper
-		h.PipelineConfig().inputsLock.Lock()
-		pw := h.PipelineConfig().inputWrappers[ir.name]
-		h.PipelineConfig().inputsLock.Unlock()
-
-		// Attempt to recreate the plugin until it works without error
-		// or until we were told to stop
-	createLoop:
-		for !globals.Stopping {
-			err := rh.Wait()
-			if err != nil {
-				ir.LogError(err)
-				globals.ShutDown()
-				return
-			}
-			p, err := pw.CreateWithError()
-			if err != nil {
-				ir.LogError(err)
-				continue
-			}
-			ir.plugin = p.(Plugin)
-			rh.Reset()
-			break createLoop
-		}
-		ir.LogMessage("exited, now restarting.")
-	}
-}
-
-func (ir *iRunner) Inject(pack *PipelinePack) {
-	ir.h.PipelineConfig().router.InChan() <- pack
-}
-
-func (ir *iRunner) LogError(err error) {
-	log.Printf("Input '%s' error: %s", ir.name, err)
-}
-
-func (ir *iRunner) LogMessage(msg string) {
-	log.Printf("Input '%s': %s", ir.name, msg)
-}
-
-// Input plugin interface type.
-type Input interface {
-	// Start listening for / gathering incoming data, populating
-	// PipelinePacks, and passing them along to a Decoder or the Router as
-	// appropriate.
-	Run(ir InputRunner, h PluginHelper) (err error)
-	// Called as a signal to the Input to stop listening for / gathering
-	// incoming data and to perform any necessary clean-up.
-	Stop()
-}
-
 // ConfigStruct for NetworkInput plugins.
 type NetworkInputConfig struct {
 	// String representation of the address of the network connection on which
